[Tracer Library] Implement Tracer
Class Tracer implements ITracer interface. It is based on concurrent dictionary
of stacks for each thread. The stacks forms the tree of methods call, the root element belongs to
class ThreadData. If unexpected state occurrs, the tracer throws 2 types of exceptions:
TracerThreadException and TracerStackException.

Returned result contains tracer internal references (may be it should be a clone?) - a
collection of trees.